refactored iterators out of the grid tests
