+++ unit_tests/src/grid_view_tests.cpp
@@ -2,7 +2,7 @@
 // Created by David Price on 02/06/2020.
 //
 
-#include "../include/catch.hpp"
+#include "catch.hpp"
 
 #include "grid.h"
 #include "grid_view.h"
